Refactor CLI to modular subcommands and entrypoint app
