+++ pyproject.toml
@@ -49,7 +49,7 @@ tests = ["pytest"]
 dev = ["lisbet[tests,docs]"]
 
 [project.scripts]
-betman = "lisbet.cli:main"
+betman = "lisbet.cli:app"
 
 [tool.pytest.ini_options]
 addopts = ["--import-mode=importlib"]
