+++ src/lisbet/cli/__init__.py
@@ -1,0 +1,9 @@
+from lisbet.cli.main import app
+
+__all__ = ["app"]
+__docs__ = """
+LISBET CLI
+==========
+Command line interface for LISBET, a framework for behavior analysis and machine
+learning.
+"""
